Node Endpoint Changes and Better Code Readability
The "Initial Connection Endpoint" functionality is now part of the "Get Nodes Endpoint".
The "Connection Check Endpoint" now optionally takes in a JSON node as "input", in order for the client(s) and/or DNS server(s) to "connect".
As a result, the client and DNS server now have the same exact endpoints.
Functions for Local and General Node-Endpoint-specific requests have been defined. Those are used in the DNS server and client to improve code readability.
Using the json_destruct_node() function of node.py its now possible to print out the IP address - port pair in the __repr__ format defined in the Node class.
The README is changed accordingly, to specify "Node Endpoints" instead of repeating the same endpoints for client / DNS server.