Tweak in Node Endpoints and Draft Blockchain Info
The DNS server was accidentally being added as a node in the clients nodes.json file. This has now been fixed, by making the JSON node in the general request helper function of the "connection check" endpoint optional. The same has also been done for the general "get nodes" function.

Draft blockchain info endpoints and the general structure has also been implemented. This change led to the creation of a new folder within the client's directory called "blocks" which will include the various blocks as individual blockX.json files. The required changes in the client settings and arguments have also been done.
For now, some draft information is also included, and one endpoint is for getting that info, and one for completely replacing it.
Let's also mention that, there are also construction, destruction and is_valid functions for the Blockchain (Info) structure.