Transaction and Block UTXO Validation
With this commit, only valid transactions will be added to blocks. More specifically, when a transaction is posted, each input will be checked using an UTXO output retrieval request. If the UTXO output truly exists then the transaction input will be valid.

In the case of forming blocks, the transactions referenced should be in the unconfirmed transactions structure (transactions.json) and not referencing the same UTXO outputs (double-spending check). Thus, the steps for validation a block are now as follows:
- recalculate and check hash
- check if all transactions are in the unconfirmed transactions structure
- check if all the inputs of the transactions are valid UTXO entries and also keep track of UTXO referencing to prevent double-spending
After that the create block callback now also updates the transactions.json file by removing all transactions that have been included in the block, and also updates the UTXOs, by creating a new entry for the block reward transaction output, removing spent UTXOs (inputs) and creating new UTXO entries for the outputs, for the remainder transactions.

The UTXO endpoints have also been tweaked in order to correctly calculate the balance when adding/removing new entries.

The new client.py testing thread now simply posts the transactions and blocks, with the UTXO being handled internally by the block creation callback.