Revamp Remaining Requests and Endpoints
The remaining requests now also use the new common/requests.py format.
All the endpoints now return 200 when the request is successful and 400 when something went wrong. Later on, we might even get more specific, but that's sufficient for now!
Let's also mention that multiple threads are spawned when making network relay requests.